holy boilerplate batman

+++ rayduino/renderer.cpp
@@ -5,5 +5,6 @@ Renderer::Renderer() {
 }
 
 Renderer::Renderer(int length, int width) {
-    projectionPlane = Vector3((float) length / 2, (float) width / 2, depth);
+    // initializes the projection plane to be in the center of the width and height
+    projectionPlane = Vector3(length / 2.0, width / 2.0, depth);
 };
